refactor(#8/post): exception 패키지 위치를 application 하위로 변경
