+++ src/main/java/org/sopt/post/application/query/PostQueryService.java
@@ -11,7 +11,7 @@ import org.sopt.post.application.dto.response.SearchResultServiceResponse;
 import org.sopt.post.domain.Post;
 import org.sopt.post.domain.Tag;
 import org.sopt.post.infrastructure.repository.PostRepository;
-import org.sopt.post.presentation.exception.PostErrorCode;
+import org.sopt.post.application.exception.PostErrorCode;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
